test: refactored to simplify and speedup
The tests were not parameterised very well, in that the same test was being run multiple times to check each field for a given test scenario. While with unit tests one assertion per test is a good rule of thumb, here the thing under test is the operation, and the setup and tear down is heavy (creating / truncating database tables, etc), so by grouping together sets of field checks, parameterisation could be reduced to different scenarios, and this has resulted in a big test speed up, as well as easier to read tests.