fix: add git to build for new suitecrm library
This commit creates a builder image into which git is installed
so that when installing dependencies iati-registry-libsuitecrm
(not on PyPI and installed direct from GitHub) can be fetched.
Dependencies now installed into a venv, which is copied to a new
image to avoid having git and other tools in the final image.